refactor(workflows): neutralize automation runtime service naming
