+++ turbo/apps/api/src/signals/services/github-webhook-automation-event.service.ts
@@ -35,7 +35,7 @@ import { workflowAutomationCanFire } from "./workflow-automation-access.service"
 import { runWorkflowAutomationNow$ } from "./workflow-automation-run.service";
 import type { AutomationRow } from "./zero-workflow-automation-launch.service";
 import type { WorkflowAutomationContext } from "./workflow-automation-context.service";
-import { ensureWorkflowUserAutomationThread } from "./zero-workflow-user-automation-thread.service";
+import { ensureWorkflowUserAutomationThread } from "./workflow-user-automation-thread.service";
 import {
   AutomationEventSourceTiming,
   type AutomationEventRunTiming,
