+++ turbo/apps/api/src/signals/services/workflow-user-automation-thread.service.ts
