+++ turbo/apps/api/src/signals/services/zero-workflow-automation.service.ts
@@ -128,7 +128,7 @@ import { manualTriggerSource } from "./workflow-automation-trigger-source";
 import {
   ensureWorkflowUserAutomationThread,
   loadWorkflowUserAutomationThreadId,
-} from "./zero-workflow-user-automation-thread.service";
+} from "./workflow-user-automation-thread.service";
 import { buildWorkflowScheduleAutomationBrief } from "./workflow-automation-brief.service";
 import type { WorkflowAutomationContext } from "./workflow-automation-context.service";
 import { reconcileAutomationEventWatches } from "./automation-event-watch-lifecycle.service";
